\name{print.comprisk.ROC}
\alias{print.comprisk.ROC}
\title{Print an object of class comprisk.ROC}

\description{
  S3 method for class ''comprisk.ROC''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{comprisk.ROC}(x, ...)
}

\arguments{
  \item{x}{output from the function comprisk.ROC}
  \item{\dots}{NA}
}
\value{
  NULL
}

